+++ examples/live/bybit/bybit_market_maker.py
@@ -88,6 +88,7 @@ config_node = TradingNodeConfig(
             product_types=[product_type],  # Will load all instruments
             demo=False,  # If client uses the demo API
             testnet=False,  # If client uses the testnet API
+            recv_window_ms=5_000,  # Default
         ),
     },
     exec_clients={
@@ -103,6 +104,7 @@ config_node = TradingNodeConfig(
             testnet=False,  # If client uses the testnet API
             max_retries=3,
             retry_delay=1.0,
+            recv_window_ms=5_000,  # Default
         ),
     },
     timeout_connection=20.0,